Bug 40097: Update toolchain for Fenix 82
As Fenix 82 uses clang 11.0.0-rc3 it is easier if we switch to using
the llvm-project repository directly. This fixes the mobile part of
bug 40048.

We need to use our compiled binutils version, too, as the one that comes
with Debian Buster (2.31.1) is breaking the build. We hit
https://bugs.llvm.org/show_bug.cgi?id=42994 which is essentially
an ld bug (https://sourceware.org/bugzilla/show_bug.cgi?id=24458). That
one is fixed on the 2.31 branch and onwards but never made it into a
2.31.x release.