Bug 40097: Update toolchain for Fenix 82
application-services needs a newer Rust compiler as it now uses
features (strip_suffix()) that got stablizied in 1.45.0 first. We don't
want to risk GeckoView stability and thus bite the bullet and add a
different project solely used by application-services for the time being.
We omit the clean-up as this is supposed to be temporary until Rust for
GeckoView gets bumped.

It is noteworthy that we still need the patch for bug 43909 albeit in
a slightly updated fashion. Originally, this landed on trunk for the
10.0 release but it got backed out later on and did not reland in time
for any 10.0 release. Thus, we need to keep that patch until Rust
switches to LLVM 11.